CU-2wgnqg5 Separate some of the regression code into different modules
